# Improve performance by removing extraneous `clone`s and tweaking initial buffer sizes.