Improve performance by removing extraneous `clone`s and tweaking initial buffer sizes.

```
10:38❯ cargo bench
    Finished bench [optimized + debuginfo] target(s) in 0.68s
     Running benches/bench_eval_multi_like.rs (target/release/deps/bench_eval_multi_like-40a62f267e10d031)
parse-1                 time:   [6.7670 µs 6.8086 µs 6.8564 µs]
                        change: [-8.6514% -7.1434% -5.6991%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  7 (7.00%) high mild
  1 (1.00%) high severe

parse-15                time:   [58.581 µs 58.882 µs 59.206 µs]
                        change: [-4.1975% -2.8944% -1.4224%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe

parse-30                time:   [118.08 µs 118.62 µs 119.16 µs]
                        change: [-4.8603% -3.3925% -2.2699%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe

compile-1               time:   [20.189 µs 20.269 µs 20.357 µs]
                        change: [+0.4576% +1.5569% +2.7740%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe

compile-15              time:   [58.767 µs 59.281 µs 59.992 µs]
                        change: [-1.3774% -0.2256% +0.9734%] (p = 0.72 > 0.05)
                        No change in performance detected.
Found 11 outliers among 100 measurements (11.00%)
  7 (7.00%) high mild
  4 (4.00%) high severe

compile-30              time:   [100.08 µs 100.74 µs 101.44 µs]
                        change: [-4.9675% -2.5715% -0.6447%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) high mild
  5 (5.00%) high severe

plan-1                  time:   [21.575 µs 21.696 µs 21.828 µs]
                        change: [-3.6081% -2.9408% -2.1841%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

plan-15                 time:   [335.78 µs 337.69 µs 339.75 µs]
                        change: [-2.9899% -2.3007% -1.6301%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

plan-30                 time:   [674.52 µs 676.98 µs 679.50 µs]
                        change: [-4.7574% -3.9435% -3.0584%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe

eval-1                  time:   [24.502 ms 24.629 ms 24.774 ms]
                        change: [-25.058% -24.374% -23.704%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

Benchmarking eval-15: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 16.6s, or reduce sample count to 30.
eval-15                 time:   [164.89 ms 165.41 ms 165.97 ms]
                        change: [-6.9389% -6.4914% -6.0866%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe

Benchmarking eval-30: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 31.9s, or reduce sample count to 10.
eval-30                 time:   [321.79 ms 322.94 ms 324.26 ms]
                        change: [-5.0670% -4.2883% -3.6095%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

     Running benches/bench_eval.rs (target/release/deps/bench_eval-314331151edf3f8b)
join                    time:   [33.455 µs 33.737 µs 34.092 µs]
                        change: [-8.5408% -7.3388% -6.3447%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  9 (9.00%) high mild
  2 (2.00%) high severe

simple                  time:   [8.3750 µs 8.4392 µs 8.5179 µs]
                        change: [-1.4552% -0.4105% +0.5757%] (p = 0.45 > 0.05)
                        No change in performance detected.
Found 13 outliers among 100 measurements (13.00%)
  3 (3.00%) high mild
  10 (10.00%) high severe

simple-no               time:   [3.8393 µs 3.8610 µs 3.8858 µs]
                        change: [-2.4937% -1.3078% -0.2708%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  9 (9.00%) high mild
  3 (3.00%) high severe

numbers                 time:   [110.93 ns 111.48 ns 112.08 ns]
                        change: [-2.0647% -0.6473% +0.5693%] (p = 0.36 > 0.05)
                        No change in performance detected.
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) high mild
  11 (11.00%) high severe

     Running benches/bench_parse.rs (target/release/deps/bench_parse-b606eba6d5f82c02)
parse-simple            time:   [1.0609 µs 1.0677 µs 1.0753 µs]
                        change: [-1.9550% -0.9580% -0.0189%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

parse-ion               time:   [3.0233 µs 3.0427 µs 3.0651 µs]
                        change: [-1.4378% -0.4134% +0.6009%] (p = 0.44 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low mild
  8 (8.00%) high mild
  3 (3.00%) high severe

parse-group             time:   [9.4457 µs 9.6779 µs 9.9555 µs]
                        change: [-0.1578% +1.2818% +2.7020%] (p = 0.09 > 0.05)
                        No change in performance detected.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

parse-complex           time:   [23.061 µs 23.302 µs 23.600 µs]
                        change: [-2.5563% -1.4074% -0.2360%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 15 outliers among 100 measurements (15.00%)
  6 (6.00%) high mild
  9 (9.00%) high severe

parse-complex-fexpr     time:   [36.844 µs 36.970 µs 37.109 µs]
                        change: [-4.0745% -3.0001% -1.9867%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
```

---

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.